[#32] refactor: 컴포넌트 위치 수정 회원가입/로그인에서 사용되는 컴포넌트들을 모아놓는 위치로 수정
